Update to per-course user API
